requirements.txt
aioredis==1.0.0
asgiref==2.3.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
certifi==2018.4.16
channels==2.1.1
channels-redis==2.1.1
chardet==3.0.4
constantly==15.1.0
daphne==2.1.1
Django==1.11.10
djangorestframework==3.8.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
requests==2.18.4
six==1.11.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
zope.interface==4.5.0
